+++ HISTORY.md
@@ -2,6 +2,13 @@
 
 - TODO: refactor, add test cases & test | update stat & go-helper functions
 
+## Version 0.1.1
+
+- Release date: Dec 08, 2020
+- Updated github package link / reference
+- Patched to use mctestgo v0.5.1, mcresponsego v0.3.1
+- Added test cases file placeholder
+
 ## Version 0.1.0
 
 - Release date: Dec 06, 2020